replaced size_t with uint64_t

+++ axhashmap.h
@@ -7,7 +7,6 @@
 
 #include <stdint.h>
 #include <stdbool.h>
-#include <stddef.h>
 
 typedef struct axhashmap axhashmap;
 
@@ -20,24 +19,24 @@ axhashmap *axh_setLoadFactor(axhashmap *h, double lf);
 
 double axh_getLoadFactor(axhashmap *h);
 
-axhashmap *axh_setDynamicSpan(axhashmap *h, size_t (*dynamicSpan)(const void *));
+axhashmap *axh_setDynamicSpan(axhashmap *h, uint64_t (*dynamicSpan)(const void *));
 
-size_t (*axh_getDynamicSpan(axhashmap *h))(const void *);
+uint64_t (*axh_getDynamicSpan(axhashmap *h))(const void *);
 
 axhashmap *axh_setDestructor(axhashmap *h, void (*destroy)(void *, void *));
 
 void (*axh_getDestructor(axhashmap *h))(void *, void *);
 
 
-axhashmap *axh_sizedNew(size_t size, size_t span, double loadFactor);
+axhashmap *axh_sizedNew(uint64_t size, uint64_t span, double loadFactor);
 
-axhashmap *axh_new(size_t span);
+axhashmap *axh_new(uint64_t span);
 
 void axh_destroy(axhashmap *h);
 
-bool axh_rehash(axhashmap *h, size_t size);
+bool axh_rehash(axhashmap *h, uint64_t size);
 
-int axh_sizedMap(axhashmap *h, void *key, void *value, size_t size);
+int axh_sizedMap(axhashmap *h, void *key, void *value, uint64_t size);
 
 int axh_map(axhashmap *h, void *key, void *value);
